requirements.txt
pandas
numpy
python-dotenv
matplotlib
sqlalchemy<2.0
mysqlclient
tqdm
python-dotenv
plotly
nbformat>=4.2.0
yfinance
streamlit
lightgbm
lightweight-charts
scikit-learn==1.3.2
joblib==1.3.2
mlflow
requests
streamlit-aggrid
ts-api